+++ src/transformers/inject-dependencies.js
@@ -1,34 +1,6 @@
 const { SourceMapGenerator } = require('source-map')
 
-const S = require('../sanctuary')
-
-const appendSemi = r => `${r};`
-
-const requireWithAssignment = (name, module, type = 'const') =>
-  `${type} ${name} = require('${module || name}')`
-
-const simpleRequire = module => `require('${module}')`
-
-const handleDep = S.ifElse(
-  S.is(String),
-  requireWithAssignment,
-  ({ name, module, type }) =>
-    name ? requireWithAssignment(name, module, type) : simpleRequire(module)
-)
-
-const assignmentExpression = (key, value) => `let ${key} = ${value}`
-
-const listDependencies = S.pipe([
-  S.map(handleDep),
-  S.map(appendSemi),
-  S.joinWith('\n'),
-])
-
-const listGlobals = globals =>
-  Object.keys(globals)
-    .map(key => assignmentExpression(key, globals[key]))
-    .map(appendSemi)
-    .join('\n')
+const { listDependencies, listGlobals } = require('../utils')
 
 const generateSourceMaps = (
   file = 'snippet.js',
